--- a/parfive/downloader.py
+++ b/parfive/downloader.py
@@ -321,7 +321,7 @@ class Downloader:
 
     @staticmethod
     async def _get_http(session, *, url, filepath_partial, chunksize=100,
-                        file_pb=None, token, overwrite, timeouts, **kwargs):
+                        file_pb=None, token, overwrite, timeouts, max_splits=12, **kwargs):
         """
         Read the file from the given url into the filename given by ``filepath_partial``.
 
@@ -346,6 +346,9 @@ class Downloader:
 
         token : `parfive.downloader.Token`
             A token for this download slot.
+
+        max_splits: `int`
+            Number of maximum concurrent connections.
 
         kwargs : `dict`
             Extra keyword arguments are passed to `aiohttp.ClientSession.get`.
@@ -372,10 +375,119 @@ class Downloader:
                                           total=get_http_size(resp))
                     else:
                         file_pb = None
-                    _write_response_to_file(resp, filepath, file_pb, chunksize)
+                    if max_splits and resp.headers.get('Accept-Ranges', None) == "bytes":
+                        # XXX: int(?)
+                        content_length = int(resp.headers['Content-length'])
+                        split_length = content_length//max_splits
+                        ranges = [
+                            [start, start + split_length]
+                            for start in range(0, content_length, split_length)
+                        ]
+                        # let the last part download everything
+                        ranges[-1][1] = ''
+                        queue = asyncio.Queue()
+                        workers = []
+                        for _range in ranges:
+                            workers.append(
+                                asyncio.create_task(Downloader._ranged_http_producer(
+                                    session, url, chunksize, _range, timeout, queue, **kwargs
+                                ))
+                            )
+                        consumer = asyncio.create_task(
+                            Downloader._ranged_http_consumer(queue, file_pb, filepath))
+                        await asyncio.gather(*workers)
+                        await queue.join()
+                        consumer.cancel()
+                        return str(filepath)
+                    else:
+                        with open(str(filepath), 'wb') as fd:
+                            while True:
+                                chunk = await resp.content.read(chunksize)
+                                if not chunk:
+                                    # Close the file progressbar
+                                    if file_pb is not None:
+                                        file_pb.close()
+
+                                    return str(filepath)
+
+                                # Write this chunk to the output file.
+                                fd.write(chunk)
+
+                                # Update the progressbar for file
+                                if file_pb is not None:
+                                    file_pb.update(chunksize)
 
         except Exception as e:
             raise FailedDownload(filepath_partial, url, e)
+
+    @staticmethod
+    async def _ranged_http_consumer(queue, file_pb, filepath):
+        """
+
+        queue: `asyncio.Queue`
+             Queue for chunks
+
+        file_pb : `tqdm.tqdm` or `False`
+            Should progress bars be displayed for each file downloaded.
+
+        filepath: `pathlib.Path`
+            Path to the which the file should be downloaded.
+        """
+        with open(filepath, 'wb') as f:
+            while True:
+                offset, chunk = await queue.get()
+                # breakpoint()
+
+                f.seek(offset)
+                f.write(chunk)
+                f.flush()
+
+                # Update the progressbar for file
+                if file_pb is not None:
+                    file_pb.update(len(chunk))
+
+                queue.task_done()
+
+    @staticmethod
+    async def _ranged_http_producer(session, url, chunksize, http_range, timeout, queue, **kwargs):
+        """
+        Worker for ranged http requests.
+
+        Parameters
+        ----------
+
+        session : `aiohttp.ClientSession`
+            The `aiohttp.ClientSession` to use to retrieve the files.
+
+        url : `str`
+            The url to retrieve.
+
+        chunksize : `int`
+            The number of bytes to read into the file at a time.
+
+        http_range: (int, int)
+             Start and end bytes of the file
+
+        queue: `asyncio.Queue`
+             Queue for chunks
+
+        kwargs : `dict`
+            Extra keyword arguments are passed to `aiohttp.ClientSession.get`.
+        """
+        headers = kwargs.pop('headers', {})
+        headers['Range'] = 'bytes={}-{}'.format(*http_range)
+
+        # init offset to start of range
+        offset, _ = http_range
+
+        async with session.get(url, timeout=timeout, headers=headers, **kwargs) as resp:
+            while True:
+                chunk = await resp.content.read(chunksize)
+                if not chunk:
+                    # print(f"{http_range[0]} done")
+                    break
+                await queue.put((offset, chunk))
+                offset += len(chunk)
 
     @staticmethod
     async def _get_ftp(session=None, *, url, filepath_partial,
@@ -447,22 +559,3 @@ class Downloader:
 
         except Exception as e:
             raise FailedDownload(filepath_partial, url, e)
-
-
-async def _write_response_to_file(resp, filepath, file_pb, chunksize):
-    with open(str(filepath), 'wb') as fd:
-        while True:
-            chunk = await resp.content.read(chunksize)
-            if not chunk:
-                # Close the file progressbar
-                if file_pb is not None:
-                    file_pb.close()
-
-                return str(filepath)
-
-            # Write this chunk to the output file.
-            fd.write(chunk)
-
-            # Update the progressbar for file
-            if file_pb is not None:
-                file_pb.update(chunksize)