--- a/parfive/utils.py
+++ b/parfive/utils.py
@@ -1,4 +1,5 @@
 import cgi
+import hashlib
 import pathlib
 from itertools import count
 
@@ -100,6 +101,19 @@ def get_filepath(filepath, overwrite):
         filepath.parent.mkdir(parents=True)
 
     return filepath, False
+
+
+def sha256sum(filename):
+    """
+    https://stackoverflow.com/a/44873382
+    """
+    h  = hashlib.sha256()
+    b  = bytearray(128*1024)
+    mv = memoryview(b)
+    with open(filename, 'rb', buffering=0) as f:
+        for n in iter(lambda : f.readinto(mv), 0):
+            h.update(mv[:n])
+    return h.hexdigest()
 
 
 class FailedDownload(Exception):