| File size | Eariler | With 12 concc conn |
| --------- | ------- | ------------------ |
| 10MB      | 10s     | 5s                 |
| 100MB     | 36s     | 21s                |