@vn-ki 2019-07-27
@Cadair I couldn't test ftp. Can you give me a server which you used for testing?