@Cadair 2019-07-27
For FTP I have just been using some of the servers sunpy uses.