@Cadair 2019-07-27
One thing I would like to check before we merge this is that the SunPy online test suite passes using it.