@vn-ki 2019-07-27
> For FTP I have just been using some of the servers sunpy uses.

I tried this. Something is up with my internet. Can you try it out for me?