@Cadair 2019-07-27
The test suite is passing for me locally:
```
 parfive/tests/test_downloader.py ✓✓✓✓✓✓✓✓✓✓✓✓✓✓✓✓✓✓✓                                                                                                                                                               100% ██████████

----------- coverage: platform linux, python 3.7.3-final-0 -----------
Name                    Stmts   Miss  Cover
-------------------------------------------
parfive/downloader.py     204     17    92%
parfive/results.py         38     10    74%
parfive/utils.py           79      8    90%
-------------------------------------------
TOTAL                     321     35    89%


Results (11.71s):
      19 passed
```