@Cadair 2019-08-04
This looks great, but needs some tests.